# openapi/productize-daemon.json
{
    "components": {
        "parameters": {
...
            "Limit": {
                "description": "Maximum number of runs to return.",
                "in": "query",
                "name": "limit",
                "schema": {
                    "default": 100,
                    "maximum": 500,
                    "minimum": 1,
                    "type": "integer"
                }
            },
...
            "RunMode": {
                "description": "Optional run mode filter.",
                "in": "query",
                "name": "mode",
                "schema": {
                    "type": "string"
                }
            },
            "RunStatus": {
                "description": "Optional run status filter.",
                "in": "query",
                "name": "status",
                "schema": {
                    "type": "string"
                }
            },
...
            "WorkspaceFilterQuery": {
                "description": "Optional workspace filter for run inventory reads.",
                "in": "query",
                "name": "workspace",
                "schema": {
                    "type": "string"
                }
            },
...
            "ActiveWorkspaceHeader": {
                "description": "Active workspace identifier for browser-scoped daemon requests.",
                "in": "header",
                "name": "X-Productize-Workspace-ID",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        },
        "responses": {
...
            "Forbidden": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Request was rejected by browser-facing transport security."
            },
            "InternalError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Unexpected daemon failure."
            },
            "InvalidRequest": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Request body or parameters were malformed."
            },
            "NotFound": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Requested resource was not found."
            },
            "StaleWorkspace": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Active workspace context is missing or stale."
            },
            "ValidationError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TransportError"
                        }
                    }
                },
                "description": "Request validation failed."
            }
        },
        "schemas": {
...
            "DaemonHealth": {
                "properties": {
                    "degraded": {
                        "type": "boolean"
                    },
                    "details": {
                        "items": {
                            "$ref": "#/components/schemas/HealthDetail"
                        },
                        "type": "array"
                    },
                    "ready": {
                        "type": "boolean"
                    }
                },
                "required": ["ready"],
                "type": "object"
            },
...
            "DaemonStatus": {
                "properties": {
                    "active_run_count": {
                        "type": "integer"
                    },
                    "http_port": {
                        "type": "integer"
                    },
                    "pid": {
                        "type": "integer"
                    },
                    "socket_path": {
                        "type": "string"
                    },
                    "started_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    },
                    "workspace_count": {
                        "type": "integer"
                    }
                },
                "required": ["active_run_count", "pid", "started_at", "workspace_count"],
                "type": "object"
            },
...
            "DashboardPayload": {
                "properties": {
                    "active_runs": {
                        "items": {
                            "$ref": "#/components/schemas/Run"
                        },
                        "type": "array"
                    },
                    "daemon": {
                        "$ref": "#/components/schemas/DaemonStatus"
                    },
                    "health": {
                        "$ref": "#/components/schemas/DaemonHealth"
                    },
                    "pending_reviews": {
                        "type": "integer"
                    },
                    "queue": {
                        "$ref": "#/components/schemas/DashboardQueueSummary"
                    },
                    "workflows": {
                        "items": {
                            "$ref": "#/components/schemas/WorkflowCard"
                        },
                        "type": "array"
                    },
                    "workspace": {
                        "$ref": "#/components/schemas/Workspace"
                    }
                },
                "required": ["daemon", "health", "pending_reviews", "queue", "workspace"],
                "type": "object"
            },
            "DashboardQueueSummary": {
                "properties": {
                    "active": {
                        "type": "integer"
                    },
                    "canceled": {
                        "type": "integer"
                    },
                    "completed": {
                        "type": "integer"
                    },
                    "failed": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "required": ["active", "canceled", "completed", "failed", "total"],
                "type": "object"
            },
            "DashboardResponse": {
                "properties": {
                    "dashboard": {
                        "$ref": "#/components/schemas/DashboardPayload"
                    }
                },
                "required": ["dashboard"],
                "type": "object"
            },
...
            "HealthDetail": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "severity": {
                        "type": "string"
                    }
                },
                "required": ["code", "message"],
                "type": "object"
            },
...
            "ReviewSummary": {
                "properties": {
                    "pr_ref": {
                        "type": "string"
                    },
                    "provider": {
                        "type": "string"
                    },
                    "resolved_count": {
                        "type": "integer"
                    },
                    "round_number": {
                        "type": "integer"
                    },
                    "unresolved_count": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "workflow_slug": {
                        "type": "string"
                    }
                },
                "required": [
                    "resolved_count",
                    "round_number",
                    "unresolved_count",
                    "updated_at",
                    "workflow_slug"
                ],
                "type": "object"
            },
...
            "Run": {
                "properties": {
                    "ended_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "error_text": {
                        "type": "string"
                    },
                    "mode": {
                        "type": "string"
                    },
                    "parent_run_id": {
                        "type": "string"
                    },
                    "presentation_mode": {
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    },
                    "run_id": {
                        "type": "string"
                    },
                    "started_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "workflow_id": {
                        "type": "string"
                    },
                    "workflow_slug": {
                        "type": "string"
                    },
                    "workspace_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "mode",
                    "presentation_mode",
                    "run_id",
                    "started_at",
                    "status",
                    "workspace_id"
                ],
                "type": "object"
            },
...
            "RunsResponse": {
                "properties": {
                    "runs": {
                        "items": {
                            "$ref": "#/components/schemas/Run"
                        },
                        "type": "array"
                    }
                },
                "required": ["runs"],
                "type": "object"
            },
...
            "SyncRequest": {
                "properties": {
                    "path": {
                        "type": "string"
                    },
                    "workflow_slug": {
                        "type": "string"
                    },
                    "workspace": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SyncResult": {
                "properties": {
                    "checkpoints_updated": {
                        "type": "integer"
                    },
                    "legacy_artifacts_removed": {
                        "type": "integer"
                    },
                    "pruned_workflows": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "review_issues_upserted": {
                        "type": "integer"
                    },
                    "review_rounds_upserted": {
                        "type": "integer"
                    },
                    "snapshots_upserted": {
                        "type": "integer"
                    },
                    "synced_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "synced_paths": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "target": {
                        "type": "string"
                    },
                    "task_items_upserted": {
                        "type": "integer"
                    },
                    "warnings": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "workflow_slug": {
                        "type": "string"
                    },
                    "workflows_scanned": {
                        "type": "integer"
                    },
                    "workflows_pruned": {
                        "type": "integer"
                    },
                    "workspace_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TransportError": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "details": {
                        "additionalProperties": true,
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    }
                },
                "required": ["request_id", "code", "message"],
                "type": "object"
            },
...
            "WorkflowCard": {
                "properties": {
                    "active_runs": {
                        "type": "integer"
                    },
                    "latest_review": {
                        "$ref": "#/components/schemas/ReviewSummary"
                    },
                    "review_round_count": {
                        "type": "integer"
                    },
                    "task_completed": {
                        "type": "integer"
                    },
                    "task_pending": {
                        "type": "integer"
                    },
                    "task_total": {
                        "type": "integer"
                    },
                    "workflow": {
                        "$ref": "#/components/schemas/WorkflowSummary"
                    }
                },
                "required": [
                    "active_runs",
                    "review_round_count",
                    "task_completed",
                    "task_pending",
                    "task_total",
                    "workflow"
                ],
                "type": "object"
            },
...
            "WorkflowSummary": {
                "properties": {
                    "archive_eligible": {
                        "type": "boolean"
                    },
                    "archive_reason": {
                        "type": "string"
                    },
                    "archived_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "can_start_run": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "string"
                    },
                    "last_synced_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "start_block_reason": {
                        "type": "string"
                    },
                    "task_counts": {
                        "$ref": "#/components/schemas/WorkflowTaskCounts"
                    },
                    "workspace_id": {
                        "type": "string"
                    }
                },
                "required": ["id", "slug", "workspace_id"],
                "type": "object"
            },
            "WorkflowTaskCounts": {
                "properties": {
                    "completed": {
                        "type": "integer"
                    },
                    "pending": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "required": ["completed", "pending", "total"],
                "type": "object"
            },
            "WorkflowsResponse": {
                "properties": {
                    "workflows": {
                        "items": {
                            "$ref": "#/components/schemas/WorkflowSummary"
                        },
                        "type": "array"
                    }
                },
                "required": ["workflows"],
                "type": "object"
            },
            "Workspace": {
                "properties": {
                    "created_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "filesystem_state": {
                        "enum": ["present", "missing"],
                        "type": "string"
                    },
                    "has_catalog_data": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "string"
                    },
                    "last_checked_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "last_sync_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "last_sync_error": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "read_only": {
                        "type": "boolean"
                    },
                    "root_dir": {
                        "type": "string"
                    },
                    "run_count": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "workflow_count": {
                        "type": "integer"
                    }
                },
                "required": [
                    "created_at",
                    "filesystem_state",
                    "has_catalog_data",
                    "id",
                    "name",
                    "read_only",
                    "root_dir",
                    "run_count",
                    "updated_at",
                    "workflow_count"
                ],
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "Productize Daemon Browser API",
        "version": "0.1.0"
    },
    "openapi": "3.0.3",
    "paths": {
...
        "/api/runs": {
            "get": {
                "operationId": "listRuns",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/WorkspaceFilterQuery"
                    },
                    {
                        "$ref": "#/components/parameters/RunStatus"
                    },
                    {
                        "$ref": "#/components/parameters/RunMode"
                    },
                    {
                        "$ref": "#/components/parameters/Limit"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RunsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "List runs across workspaces or for one workspace.",
                "tags": ["runs"]
            }
        },
...
        "/api/sync": {
            "post": {
                "operationId": "syncWorkflow",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SyncRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SyncResult"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidRequest"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "Run explicit daemon reconciliation for a workspace or workflow.",
                "tags": ["workflows"]
            }
        },
        "/api/tasks": {
            "get": {
                "operationId": "listWorkflows",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkflowsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "List workflow summaries for the active workspace.",
                "tags": ["workflows"]
            }
        },
...
        "/api/ui/dashboard": {
            "get": {
                "operationId": "getDashboard",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ActiveWorkspaceHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "412": {
                        "$ref": "#/components/responses/StaleWorkspace"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                },
                "summary": "Read the active-workspace dashboard aggregate payload.",
                "tags": ["dashboard"]
            }
        },
...
    },
...
}